Implement PodReplacementPolicy for Deployments
Consider the TerminationStarted and TerminationComplete policies when
rolling out a new deployment with a Recreate or RollingUpdate strategy.

Also consider these policies when scaling any deployment, either during or
after a rollout has been completed. Special care has to be taken when scaling
up multiple replicasets proportionally. If a replicaset cannot be fully scaled,
"deployment.kubernetes.io/replicaset-replicas-before-scale" annotation is added
to the replicaset so that it can be scaled again in the future once the extra
pods terminate.